Refactored theme files
