+++ lib/employee/hr_screens/attendance_report/daily_report.dart
@@ -1,6 +1,6 @@
 import 'package:date_picker_timeline/date_picker_timeline.dart';
 import 'package:echno_attendance/constants/colors_string.dart';
-import 'package:echno_attendance/global_theme/text_style.dart';
+import 'package:echno_attendance/global_theme/text_theme.dart';
 import 'package:echno_attendance/employee/widgets/attcard_daily.dart';
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
